+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Rename DBpedia `redirection_dump` dataset to `entity_redirections`.
+- DBpedia `entity_labels` dataset now use the common class `kgdata.models.entity.EntityLabel` to make it similar to Wikidata.
 
 ## [2.3.3](https://github.com/binh-vu/kgdata/tree/2.3.3) (2022-07-06)